+++ projects/np-ui-lib/src/public-api.ts
@@ -153,9 +153,9 @@ export * from "./lib/np-virtual-scroll/np-virtual-scroll.component";
 
 /*Dialog */
 export * from "./lib/np-dialog/np-dialog.module";
-export * from "./lib/np-dialog/np-dialog.component";
 export * from "./lib/np-dialog/np-dialog.config";
 export * from "./lib/np-dialog/np-dialog.service";
+export * from "./lib/np-dialog/np-dialog.component";
 export * from "./lib/np-dialog/np-dialog-ref";
 
 /*Steps */